+++ src/main/java/faang/school/achievement/config/AsyncConfig.java
@@ -1,5 +1,6 @@
 package faang.school.achievement.config;
 
+import faang.school.achievement.config.threadpool.WriterAchievementThreadPoolConfig;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,6 +14,7 @@ import java.util.concurrent.Executor;
 @RequiredArgsConstructor
 public class AsyncConfig {
     private final EventHandlerThreadPoolConfig threadPoolConfig;
+    private final WriterAchievementThreadPoolConfig writerAchievementThreadPoolConfig;
 
     @Bean
     public Executor eventHandlerExecutor() {
@@ -22,6 +24,18 @@ public class AsyncConfig {
         executor.setCorePoolSize(threadPoolConfig.getCorePoolSize());
         executor.setThreadNamePrefix(threadPoolConfig.getThreadNamePrefix());
         executor.initialize();
+        return executor;
+    }
+
+    @Bean
+    public Executor writerAchievementExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+
+        executor.setQueueCapacity(writerAchievementThreadPoolConfig.getQueueCapacity());
+        executor.setMaxPoolSize(writerAchievementThreadPoolConfig.getMaxPoolSize());
+        executor.setCorePoolSize(writerAchievementThreadPoolConfig.getCorePoolSize());
+        executor.initialize();
+
         return executor;
     }
 }
