+++ src/main/java/faang/school/achievement/messaging/postevent/PostEventListener.java
@@ -4,6 +4,7 @@ import faang.school.achievement.dto.post.PostEvent;
 import faang.school.achievement.eventhandler.EventHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
