+++ src/main/resources/application.yaml
@@ -30,19 +30,14 @@ spring:
     bootstrap-servers: localhost:9092
     consumer:
       group-id: group
+    channels:
+
 
 writer-achievement:
   thread-pool:
     queueCapacity: 100
     maxPoolSize: 4
     corePoolSize: 4
-
-achievements:
-  writer_achievement:
-    name: "Писатель"
-    target: 100
-
-
 
 server:
   port: 8085
@@ -63,3 +58,6 @@ event-handler-thread-pool:
 
 achievements:
   celebrity_achievement: "Знаменитость"
+  writer_achievement:
+    name: "Писатель"
+    target: 100
