+++ src/test/java/faang/school/achievement/messaging/postevent/PostEventListenerTest.java
@@ -32,6 +32,7 @@ import static org.awaitility.Awaitility.await;
 
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@Disabled
 class PostEventListenerTest {
 
     static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:15-alpine")
